# docs/openapi/ironwall-api.yaml
openapi: "3.1.0"
info:
  title: AXELUS-WAF API
  version: "1.0.0"
  description: |
    # AXELUS-WAF REST API

    Complete API reference for the AXELUS-WAF platform by **OPTIMIUM NEXUS LLC**.

    ## Authentication

    Most endpoints require one of:
    - **Bearer Token** (JWT): `Authorization: Bearer <token>`
    - **API Key**: `X-Admin-Key: <key>` (admin endpoints) or `X-API-Key: <key>` (portal endpoints)

    ## Base URLs

    | Environment | URL |
    |------------|-----|
    | Management API | `https://YOUR_WAF_HOST:9443` |
    | License Server | `http://YOUR_WAF_HOST:8090` |
    | Customer Portal | `http://YOUR_WAF_HOST:8080` |
    | Admin Center | `http://YOUR_WAF_HOST:8085` |

    ## Rate Limits

    All API endpoints are rate-limited. Responses include:
    - `X-RateLimit-Limit`: Request limit per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  contact:
    name: OPTIMIUM NEXUS LLC
    url: https://www.optimiumnexus.com
    email: contact@optimiumnexus.com
  license:
    name: Proprietary
    url: https://www.optimiumnexus.com/license
servers:
  - url: https://{host}:9443
    description: Management API
    variables:
      host:
        default: localhost
        description: Your AXELUS server hostname

  - url: http://{host}:8090
    description: License Server API
    variables:
      host:
        default: localhost

  - url: http://{host}:8080
    description: Customer Portal API
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key for license and admin endpoints

    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login

    PortalAPIKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Customer portal API key

  # ── Schemas ───────────────────────────────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:  { type: string, description: "Human-readable error message" }
        code:   { type: string, description: "Machine-readable error code" }
        details: { type: object }
    ValidationResult:
      type: object
      properties:
        valid:           { type: boolean }
        tier:            { type: string, enum: [COMMUNITY, PROFESSIONAL, ENTERPRISE, ULTIMATE, TRIAL, DEVELOPER] }
        licensee:        { type: string }
        email:           { type: string, format: email }
        key:             { type: string, example: "AX-ENT-A3F2-B9K1-M7X4-Z2P8" }
        is_lifetime:     { type: boolean }
        days_remaining:  { type: integer, description: "-1 = lifetime" }
        expires_at:      { type: string, format: date-time, nullable: true }
        features:        { type: array, items: { type: string } }
        extra_features:  { type: array, items: { type: string } }
        limits:
          $ref: '#/components/schemas/LicenseLimits'
        errors:          { type: array, items: { type: string } }
        warnings:        { type: array, items: { type: string } }
        feature_available: { type: boolean, nullable: true, description: "Only set when feature parameter provided" }
    LicenseLimits:
      type: object
      properties:
        max_sites:             { type: integer }
        max_requests_per_sec:  { type: integer }
        max_nodes:             { type: integer }
        max_users:             { type: integer }
        max_custom_rules:      { type: integer }
        support_level:         { type: string, enum: [community, email, priority, dedicated] }
        update_channel:        { type: string, enum: [stable, lts, edge] }
    LicenseResponse:
      type: object
      properties:
        id:            { type: string, format: uuid }
        key:           { type: string, example: "AX-ENT-A3F2-B9K1-M7X4-Z2P8" }
        tier:          { type: string }
        licensee:      { type: string }
        email:         { type: string }
        domain:        { type: string, nullable: true }
        issued_at:     { type: string, format: date-time }
        expires_at:    { type: string, format: date-time, nullable: true }
        is_lifetime:   { type: boolean }
        is_revoked:    { type: boolean }
        days_remaining: { type: integer }
        features:      { type: array, items: { type: string } }
        limits:
          $ref: '#/components/schemas/LicenseLimits'
        status:        { type: string }
        license_file:  { type: string, description: "PEM-encoded license (only on creation)" }
    IssueLicenseRequest:
      type: object
      required: [tier, licensee, email]
      properties:
        tier:
          type: string
          enum: [COMMUNITY, PROFESSIONAL, ENTERPRISE, ULTIMATE, TRIAL, DEVELOPER]
        licensee:        { type: string, example: "Acme Corporation" }
        email:           { type: string, format: email }
        domain:          { type: string, description: "Optional domain binding" }
        hardware_id:     { type: string, description: "Optional hardware fingerprint binding" }
        valid_for_days:  { type: integer, default: 365, description: "0 = lifetime" }
        extra_features:  { type: array, items: { type: string } }
        disabled_features: { type: array, items: { type: string } }
        custom_limits:
          $ref: '#/components/schemas/LicenseLimits'
        notes:           { type: string }
...
    Pagination:
      type: object
      properties:
        total:     { type: integer }
        page:      { type: integer }
        page_size: { type: integer }
        pages:     { type: integer }

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/licensing/validate:
    post:
      tags: [License — Public]
      summary: Validate a license
      description: |
        Validates a license file or key. Optionally checks if a specific feature is available.
        This endpoint is public and can be called without authentication.
      operationId: validateLicense
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                license_file:
                  type: string
                  description: PEM-encoded license file content
                  example: "-----BEGIN IRONWALL LICENSE-----\n...\n-----END IRONWALL LICENSE-----"
                license_key:
                  type: string
                  description: Human-readable license key
                  example: "AX-ENT-A3F2-B9K1-M7X4-Z2P8"
                feature:
                  type: string
                  description: Optional feature to check availability for
                  example: "geoip_blocking"
            examples:
              validate_by_file:
                summary: Validate by PEM file
                value:
                  license_file: "-----BEGIN IRONWALL LICENSE-----\neyJpZ...\n-----END IRONWALL LICENSE-----"
              validate_with_feature:
                summary: Validate and check feature
                value:
                  license_file: "-----BEGIN IRONWALL LICENSE-----\neyJpZ...\n-----END IRONWALL LICENSE-----"
                  feature: "ai_threat_hunting"
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '403':
          description: License is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'

  # ── LICENSE — ADMIN ──────────────────────────────────────────────────────────
  /api/v1/licensing/admin/licenses:
    get:
      tags: [License — Admin]
      summary: List all licenses
      operationId: listLicenses
      security:
        - AdminKey: []
      parameters:
        - name: tier
          in: query
          schema: { type: string, enum: [COMMUNITY, PROFESSIONAL, ENTERPRISE, ULTIMATE, TRIAL, DEVELOPER] }
        - name: revoked
          in: query
          schema: { type: boolean, default: false }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: License list with pagination
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      licenses:
                        type: array
                        items:
                          $ref: '#/components/schemas/LicenseResponse'
        '401':
          description: Missing or invalid admin key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    post:
      tags: [License — Admin]
      summary: Issue a new license
      operationId: issueLicense
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueLicenseRequest'
            examples:
              enterprise_365:
                summary: Enterprise license, 1 year
                value:
                  tier: ENTERPRISE
                  licensee: "Acme Corporation"
                  email: "admin@acme.com"
                  domain: "acme.com"
                  valid_for_days: 365
              ultimate_lifetime:
                summary: Ultimate lifetime license
                value:
                  tier: ULTIMATE
                  licensee: "BigCo International"
                  email: "cto@bigco.com"
                  valid_for_days: 0
      responses:
        '201':
          description: License issued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseResponse'
  /api/v1/licensing/admin/licenses/{id}:
    get:
      tags: [License — Admin]
      summary: Get a specific license
      operationId: getLicense
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: License details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseResponse'
        '404':
          description: License not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/licensing/admin/licenses/{id}/revoke:
    delete:
      tags: [License — Admin]
      summary: Revoke a license
      operationId: revokeLicense
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, example: "Payment dispute" }
      responses:
        '200':
          description: License revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  id:      { type: string }
  /api/v1/licensing/admin/licenses/{id}/download:
    get:
      tags: [License — Admin]
      summary: Download license file
      operationId: downloadLicense
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: License file download
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
  /api/v1/licensing/admin/licenses/bulk:
    post:
      tags: [License — Admin]
      summary: Bulk issue licenses (up to 100)
      operationId: bulkIssueLicenses
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 100
              items:
                $ref: '#/components/schemas/IssueLicenseRequest'
      responses:
        '201':
          description: Batch results
          content:
            application/json:
              schema:
                type: object
                properties:
                  issued:   { type: integer }
                  errors:   { type: array, items: { type: string } }
                  licenses: { type: array, items: { $ref: '#/components/schemas/LicenseResponse' } }
...